Remove old CFP and talk voting.
