+++ conference/settings.py
@@ -35,12 +35,6 @@ import os
 # cart.
 MAX_TICKETS = 6
 
-# URL that will receive a user who tries to access the paper submission when the CFP is closed
-# None if a 404 is returned
-CFP_CLOSED = getattr(settings, 'CONFERENCE_CFP_CLOSED', None)
-
-VOTING_CLOSED = getattr(settings, 'CONFERENCE_VOTING_CLOSED', None)
-
 VOTING_OPENED = getattr(settings, 'CONFERENCE_VOTING_OPENED', lambda conf, user: conf.voting())
 
 # Callable to check whether the user passed may vote
