+++ conference/views/__init__.py
@@ -1,14 +1,9 @@
 import os.path
-import random
-from datetime import date
-from decimal import Decimal
 
-from django import forms, http
+from django import http
 from django.conf import settings as dsettings
 
-from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.contrib.contenttypes.models import ContentType
 from django.core.urlresolvers import reverse
 from django.shortcuts import get_object_or_404, redirect, render, render_to_response
 from django.template.response import TemplateResponse
@@ -240,8 +235,7 @@ def schedule_events_booking_status(request, conference):
         else:
             v['user'] = False
         del v['booked']
-    return dat
-    a
+    return data
 
 @render_to_template('conference/schedule.xml')
 def schedule_xml(request, conference, slug):
@@ -263,259 +257,6 @@ def sponsor_json(request, sponsor):
         'slug': sponsor.slug,
         'url': sponsor.url
     }
-
-
-@login_required
-#@transaction.atomic
-def paper_submission(request):
-    try:
-        speaker = request.user.speaker
-    except models.Speaker.DoesNotExist:
-        speaker = None
-
-    conf = models.Conference.objects.current()
-
-    if not conf.cfp_start or not conf.cfp_end:
-        return TemplateResponse(request,
-                                "conference/cfp/unkown_cfp_status.html")
-
-    if date.today() < conf.cfp_start:
-        return TemplateResponse(request,
-                                "conference/cfp/cfp_not_started.html")
-
-    if date.today() > conf.cfp_end:
-        return TemplateResponse(request,
-                                "conference/cfp/cfp_already_closed.html")
-
-
-    if speaker:
-        proposed = list(speaker.talk_set.proposed(conference=settings.CONFERENCE))
-    else:
-        proposed = []
-    if not proposed:
-        fc = utils.dotted_import(settings.FORMS['PaperSubmission'])
-        form = fc(user=request.user, data=request.POST, files=request.FILES)
-    else:
-        fc = utils.dotted_import(settings.FORMS['AdditionalPaperSubmission'])
-        form = fc(data=request.POST, files=request.FILES)
-
-    if request.method == 'POST':
-        if not proposed:
-            form = fc(user=request.user, data=request.POST, files=request.FILES)
-        else:
-            form = fc(data=request.POST, files=request.FILES)
-
-        if form.is_valid():
-            if not proposed:
-                talk = form.save()
-                speaker = request.user.speaker
-            else:
-                talk = form.save(speaker=speaker)
-            messages.info(request, 'Your talk has been submitted, thank you!')
-            return redirect(reverse('cfp-thank-you-for-proposal'))
-    else:
-        if not proposed:
-            form = fc(user=request.user)
-        else:
-            form = fc()
-
-    return render(request, 'conference/paper_submission.html', {
-        'speaker': speaker,
-        'form': form,
-        'proposed_talks': proposed,
-    })
-
-
-def cfp_thank_you_for_proposal(request):
-    return TemplateResponse(
-        request, "conference/cfp/thank_you_for_proposal.html"
-    )
-
-
-def get_data_for_context(request):
-    conf = models.Conference.objects.current()
-    voting_allowed = settings.VOTING_ALLOWED(request.user)
-    talks = models.Talk.objects.proposed(conference=conf.code)
-    return conf, talks, voting_allowed
-
-
-def voting(request):
-
-    conf, talks, voting_allowed = get_data_for_context(request)
-
-    if not settings.VOTING_OPENED(conf, request.user):
-        if settings.VOTING_CLOSED:
-            return redirect(settings.VOTING_CLOSED)
-        else:
-            raise http.Http404()
-
-    if request.method == 'POST':
-        if not voting_allowed:
-            return http.HttpResponseBadRequest('anonymous user not allowed')
-
-        data = dict((x.id, x) for x in talks)
-        for k, v in [x for x in request.POST.items() if x[0].startswith('vote-')]:
-            try:
-                talk = data[int(k[5:])]
-            except KeyError:
-                return http.HttpResponseBadRequest('invalid talk')
-            except ValueError:
-                return http.HttpResponseBadRequest('id malformed')
-            if not v:
-                models.VotoTalk.objects.filter(user=request.user, talk=talk).delete()
-            else:
-                try:
-                    vote = Decimal(v)
-                except ValueError:
-                    return http.HttpResponseBadRequest('vote malformed')
-                try:
-                    o = models.VotoTalk.objects.get(user=request.user, talk=talk)
-                except models.VotoTalk.DoesNotExist:
-                    o = models.VotoTalk(user=request.user, talk=talk)
-                if not vote:
-                    if o.id:
-                        o.delete()
-                else:
-                    o.vote = vote
-                    o.save()
-        if request.is_ajax():
-            return http.HttpResponse('')
-        else:
-            return HttpResponseRedirectSeeOther(reverse('conference-voting') + '?' + request.GET.urlencode())
-    else:
-        from conference.forms import TagField, ReadonlyTagWidget, PseudoRadioSelectWidget
-        class OptionForm(forms.Form):
-            abstracts = forms.ChoiceField(
-                choices=(('not-voted', 'Not yet voted'),
-                         ('all', 'All'),
-                         ),
-                required=False,
-                initial='not-voted',
-                widget=PseudoRadioSelectWidget(),
-            )
-            talk_type = forms.ChoiceField(
-                label='Session type',
-                choices=(('all', 'All'),) + tuple(settings.TALK_TYPES_TO_BE_VOTED),
-                required=False,
-                initial='all',
-                widget=PseudoRadioSelectWidget(),
-            )
-            language = forms.ChoiceField(
-                choices=(('all', 'All'),) + tuple(settings.TALK_SUBMISSION_LANGUAGES),
-                required=False,
-                initial='all',
-                widget=PseudoRadioSelectWidget(),
-            )
-            order = forms.ChoiceField(
-                choices=(('random', 'Random order'),
-                         ('vote', 'Vote'),
-                         ('speaker', 'Speaker name'),
-                         ),
-                required=False,
-                initial='random',
-                widget=PseudoRadioSelectWidget(),
-            )
-            tags = TagField(
-                required=False,
-                widget=ReadonlyTagWidget(),
-            )
-
-        # I want to associate with each talk a "unique" number to display next to the title to be able to easily find.
-        ordinal = dict()
-        for ix, t in enumerate(talks.order_by('created').values_list('id', flat=True)):
-            ordinal[t] = ix
-
-        user_votes = models.VotoTalk.objects.filter(user=request.user.id)
-
-        # Start by sorting talks by name
-        talks = talks.order_by('speakers__user__first_name',
-                               'speakers__user__last_name')
-
-        if request.GET:
-            form = OptionForm(data=request.GET)
-            form.is_valid()
-            options = form.cleaned_data
-        else:
-            form = OptionForm()
-            options = {
-                'abstracts': 'not-voted',
-                'talk_type': 'all',
-                'language': 'all',
-                'tags': '',
-                'order': 'random',
-            }
-        # if options['abstracts'] == 'not-voted':
-        #     talks = talks.exclude(id__in=user_votes.values('talk_id'))
-        if options['talk_type'] in (tchar
-                                    for (tchar, tdef) in settings.TALK_TYPES_TO_BE_VOTED):
-            talks = talks.filter(type__startswith=options['talk_type'])
-
-        if options['language'] in (lcode
-                                   for (lcode, ldef) in settings.TALK_SUBMISSION_LANGUAGES):
-            talks = talks.filter(language=options['language'])
-
-        if options['tags']:
-            # if options['tags'] ends us a tag not associated with any talk I results
-            # in a query that results from scratch; to avoid this limit the usable tag
-            # as a filter to those associated with talk.
-            allowed = set()
-            ctt = ContentType.objects.get_for_model(models.Talk)
-            for t, usage in dataaccess.tags().items():
-                for cid, oid in usage:
-                    if cid == ctt.id:
-                        allowed.add(t.name)
-                        break
-            tags = set(options['tags']) & allowed
-            if tags:
-                talks = talks.filter(id__in=models.ConferenceTaggedItem.objects\
-                    .filter(
-                        content_type__app_label='conference', content_type__model='talk',
-                        tag__name__in=tags)\
-                    .values('object_id')
-                )
-
-        talk_order = options['order']
-        votes = dict((x.talk_id, x) for x in user_votes)
-
-        # As talks are sorted by a model connected via a m2m can I have repeated the talk, and
-        # distinct does not apply in these case.
-        #
-        # It can only filtered in python, at this point I take this opportunity to engage
-        # votes user using a single loop.
-        dups = set()
-        def filter_vote(t):
-            if t['id'] in dups:
-                return False
-            dups.add(t['id'])
-            t['user_vote'] = votes.get(t['id'])
-            t['ordinal'] = ordinal[t['id']]
-            return True
-        talks = list(filter(filter_vote, talks.values('id')))
-
-        # Fix talk order, if necessary
-        if talk_order == 'vote':
-            def key(x):
-                if x['user_vote']:
-                    return x['user_vote'].vote
-                else:
-                    return Decimal('-99.99')
-            talks = reversed(sorted(reversed(talks), key=key))
-        elif talk_order == 'random':
-            random.shuffle(talks)
-        elif talk_order == 'speaker':
-            # Already sorted
-            pass
-
-        ctx = {
-            'voting_allowed': voting_allowed,
-            'talks': list(talks),
-            'form': form,
-        }
-        if request.is_ajax():
-            tpl = 'conference/ajax/voting.html'
-        else:
-            tpl = 'conference/voting.html'
-        return render(request, tpl, ctx)
 
 
 @profile_access
