--- templates/conference/cfp/cfp_not_started.html
@@ -1,12 +1,0 @@
-{% extends "p3/base.html" %}
-
-{% block MAIN_CONTENT %}
-  <section>
-  <div class="grid-container">
-    <div class="grid-100">
-      <h1>CFP has not started yet. Watch this space.</h1>
-      <a href='/'>Go back to homepage</a>
-    </div>
-  </div>
-  </section>
-{% endblock MAIN_CONTENT %}
