--- templates/conference/cfp/unkown_cfp_status.html
@@ -1,12 +1,0 @@
-{% extends "p3/base.html" %}
-
-{% block MAIN_CONTENT %}
-  <section>
-  <div class="grid-container">
-    <div class="grid-100">
-      <h1>Unkown CFP Status.</h1>
-      <a href='/'>Go back to homepage</a>
-    </div>
-  </div>
-  </section>
-{% endblock MAIN_CONTENT %}
