+++ tests/test_cfp.py
@@ -1,592 +1,563 @@
-# coding: utf-8
-
 from datetime import timedelta
-from http.client import (
-    OK as HTTP_OK_200,
-    # NOT_FOUND as HTTP_NOT_FOUND_404,
-    FOUND as HTTP_REDIRECT_302
-)
 
-import pytest
-from django.test import TestCase
-from django.core.urlresolvers import reverse
-from django.core import mail
+from pytest import mark
 from django.conf import settings
+from django.core.urlresolvers import reverse
 from django.utils import timezone
-from django.contrib.auth.models import User
 
-from django_factory_boy import auth as auth_factories
-
-from tests.factories import AssopyUserFactory
-from conference.models import (
-    Conference,
-    Talk,
-    Speaker,
-    ConferenceTag,
-    TALK_LEVEL
+from conference.models import Conference, Talk, TALK_TYPE_CHOICES, TALK_LEVEL
+from conference.cfp import (
+    dump_relevant_talk_information_to_dict,
+    AddSpeakerToTalkForm,
 )
-from taggit.models import Tag
+
+from tests.common_tools import redirects_to, template_used
+from tests.factories import TalkFactory
+
+pytestmark = mark.django_db
 
 
-# TODO(artcz) add tests for CMS-based CFP page that redirects to the submission
-# page tested below
-
-
-# Using django's TestCase because of assertRedirects
-# however we still run it using pytest
-@pytest.mark.skip(
-    "There's a new implementation of cfp, and this one no longer needs to be"
-    " fully supported. Left as documentation. "
-    "To be removed after it's reviewed in #966"
+@mark.parametrize(
+    "url",
+    [
+        reverse("cfp:step1_submit_proposal"),
+        # using some random uuid because we just need to resolve url
+        reverse("cfp:step2_add_speakers", args=["ABCDEFI"]),
+        reverse("cfp:step3_thanks", args=["ABCDEFI"]),
+        reverse("cfp:update", args=["ABCDEFI"]),
+        reverse("cfp:update_speakers", args=["ABCDEFI"]),
+    ],
 )
-class TestCFP(TestCase):
+def test_if_cfp_pages_are_login_required(client, url):
+    response = client.get(url)
+
+    assert response.status_code == 302
+    assert redirects_to(response, "/accounts/login/")
+
+
+def test_if_cfp_pages_are_unavailable_if_cfp_is_undefined(user_client):
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+    )
+    step1_url = reverse("cfp:step1_submit_proposal")
+
+    response = user_client.get(step1_url)
+    assert response.status_code == 200
+    assert template_used(response, "ep19/bs/cfp/cfp_is_closed.html")
+
+
+def test_if_cfp_pages_are_unavailable_if_cfp_is_in_the_future(user_client):
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() + timedelta(days=2),
+    )
+    step1_url = reverse("cfp:step1_submit_proposal")
+
+    response = user_client.get(step1_url)
+    assert response.status_code == 200
+    assert template_used(response, "ep19/bs/cfp/cfp_is_closed.html")
+
+
+def test_if_cfp_pages_are_unavailable_if_cfp_is_in_the_past(user_client):
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() - timedelta(days=1),
+    )
+    step1_url = reverse("cfp:step1_submit_proposal")
+
+    response = user_client.get(step1_url)
+    assert response.status_code == 200
+    assert template_used(response, "ep19/bs/cfp/cfp_is_closed.html")
+
+
+def test_if_cfp_pages_are_available_if_cfp_is_active(user_client):
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() + timedelta(days=1),
+    )
+    step1_url = reverse("cfp:step1_submit_proposal")
+
+    response = user_client.get(step1_url)
+    assert response.status_code == 200
+    assert template_used(response, "ep19/bs/cfp/step1_talk_info.html")
+
+
+def test_validation_errors_are_handled_on_step1(user_client):
     """
-    Test everything related to CFP
+    NOTE(artcz)
+    This test is basically a placholder to get a proper branch coverage.
+    We should also (separately?) test if the validation itslef is correct.
     """
-
-    def setUp(self):
-        self.conference = Conference.objects.create(
-            code=settings.CONFERENCE_CONFERENCE,
-            name=settings.CONFERENCE_CONFERENCE,
-            # by default start with open CFP
-            cfp_start=timezone.now() - timedelta(days=2),
-            cfp_end=timezone.now()   + timedelta(days=2)
-        )
-
-        # default password is 'password123' per django_factory_boy
-        admin = User.objects.create_superuser(
-            'admin', 'admin@example.com', 'admin')
-        AssopyUserFactory(user=admin)
-
-        self.user = auth_factories.UserFactory(
-            email='joedoe@example.com', is_active=True)
-        AssopyUserFactory(user=self.user)
-
-        self.form_url = reverse("conference-paper-submission")
-
-    def test_login_required(self):
-        login_url = reverse('accounts:login') + '?next=' + self.form_url
-        response = self.client.get(self.form_url)
-        assert response.status_code == HTTP_REDIRECT_302
-        self.assertRedirects(response, login_url)
-
-    def test_accessing_cfp_form_before_CFP_is_opened(self):
-        self.conference.cfp_start = timezone.now() + timedelta(days=2)
-        self.conference.save()
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        response = self.client.get(self.form_url, follow=True)
-        assert response.status_code == HTTP_OK_200
-        self.assertTemplateUsed(response,
-                                "conference/cfp/cfp_not_started.html")
-
-    def test_accessing_cfp_form_after_CFP_is_closed(self):
-        self.conference.cfp_end = timezone.now() - timedelta(days=1)
-        self.conference.save()
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        response = self.client.get(self.form_url, follow=True)
-        assert response.status_code == HTTP_OK_200
-        self.assertTemplateUsed(response,
-                                "conference/cfp/cfp_already_closed.html")
-
-    def test_accessing_cfp_form_while_cfp_is_live(self):
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        response = self.client.get(self.form_url, follow=True)
-        assert response.status_code == HTTP_OK_200
-        self.assertTemplateUsed(response, "conference/paper_submission.html")
-
-    def test_post_some_proposals(self):
-        assert Talk.objects.all().count() == 0
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        assert Speaker.objects.count() == 0
-        with self.assertRaises(Speaker.DoesNotExist):
-            self.user.speaker
-
-        VALIDATION_FAILED_200     = HTTP_OK_200
-        VALIDATION_SUCCESSFUL_302 = 302
-
-        required_proposal_fields = [
-            'title', 'abstract', 'abstract_short', 'level', 'tags',
-            'slides_agreement', 'video_agreement', 'type',
-        ]
-
-        required_speaker_fields = [
-            'bio', 'first_name', 'last_name', 'phone',
-            'birthday', 'personal_agreement',
-        ]
-
-        required_fields_for_first_proposal =\
-            required_speaker_fields + required_proposal_fields
-        required_fields_for_next_proposal = required_proposal_fields
-
-        def required_fields_not_filled(response, fields):
-            """
-            Helper function
-            """
-            errors = response.context['form'].errors
-
-            for f in fields:
-                assert errors[f] == ['This field is required.']
-                del errors[f]
-
-            # we make sure there are no other errors in the form
-            assert errors == {}
-
-        empty_submission = {}
-        response = self.client.post(self.form_url, empty_submission)
-        assert response.status_code == VALIDATION_FAILED_200
-        required_fields_not_filled(response,
-                                   required_fields_for_first_proposal)
-
-        dummy_data_in_all_required_fields = {
-            f: "foo"
-            for f in required_fields_for_first_proposal
-        }
-        response = self.client.post(self.form_url,
-                                    dummy_data_in_all_required_fields)
-        assert response.status_code == VALIDATION_FAILED_200
-
-        # birthday, type and level require specific data
-        assert {'birthday', 'type', 'level'} ==\
-            set(response.context['form'].errors.keys())
-
-        talk_proposal = {
-            "type": "t_30",
-            'first_name': 'Joe',
-            'last_name': 'Doe',
-            "birthday": "2018-02-26",
-            'bio': "Python developer",
-            "title": "Testing EPCON CFP",
-            "abstract_short": "Short talk about testing CFP",
-            "abstract": "Using django TestCase and pytest",
-            "level": "advanced",
-            "phone": "41331237",
-            "tags": "django, testing, slides",
-            "personal_agreement": True,
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        profile_url = reverse("conference-myself-profile")
-        thank_you_url = reverse('cfp-thank-you-for-proposal')
-        response = self.client.post(self.form_url, talk_proposal)
-        self.assertRedirects(response, thank_you_url,
-                             status_code=VALIDATION_SUCCESSFUL_302,
-                             fetch_redirect_response=False)
-        talk_url = Talk.objects.get().get_absolute_url()
-        assert talk_url == "/conference/talks/testing-epcon-cfp"
-
-        # At last check if the talk is visible in the list of submitted talks
-        response = self.client.get(profile_url, follow=True)
-        assert response.status_code == HTTP_OK_200
-        self.assertTemplateUsed(response, 'conference/profile.html')
-        self.assertContains(response, talk_proposal['title'])
-        self.assertContains(response, talk_url)
-
-        # Now that a talk is already proposed, second proposal will look
-        # slightly different
-        assert Speaker.objects.get()
-        response = self.client.get(self.form_url)
-        self.assertContains(response, "You have already submitted 1 proposal")
-
-        response = self.client.post(self.form_url, empty_submission)
-        required_fields_not_filled(response, required_fields_for_next_proposal)
-
-        talk_proposal = {
-            "type": "t_45",
-            "title": "More about EPCON testing",
-            "abstract_short": "Longer talk about testing",
-            "abstract": "Using django TestCase and pytest",
-            "level": "advanced",
-            "tags": "django, testing, slides",
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-        assert Talk.objects.all().count() == 2
-
-        # check the form again
-        response = self.client.get(self.form_url)
-        self.assertContains(response, "You have already submitted 2 proposals")
-
-        # NOTE(artcz)(2018-02-27)
-        # This assertion checks if we're not sending emails during the
-        # submission process.
-        # There is a signal that is triggered when new Talk is created
-        # (_new_paper_email) and it sends emails to people on the
-        # CONFERENCE_TALK_SUBMISSION_NOTIFICATION_EMAIL list.
-        # For now this list is empty, so it shouldn't send any emails. If we
-        # ever change that (either by removing the signal or putting some
-        # emails on that list) – we should update this test
-        assert len(mail.outbox) == 0
-
-    def test_ignores_new_tags(self):
-        assert Talk.objects.all().count() == 0
-        assert ConferenceTag.objects.count() == 0
-        assert Tag.objects.count() == 0
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        VALIDATION_SUCCESSFUL_302 = 302
-
-        talk_proposal = {
-            "type": "t_30",
-            'first_name': 'Joe',
-            'last_name': 'Doe',
-            "birthday": "2018-02-26",
-            'bio': "Python developer",
-            "title": "Testing EPCON CFP",
-            "abstract_short": "Short talk about testing CFP",
-            "abstract": "Using django TestCase and pytest",
-            "level": "advanced",
-            "phone": "41331237",
-            "tags": "django, testing, slides",
-            "personal_agreement": True,
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        assert ConferenceTag.objects.count() == 0
-        assert Tag.objects.count() == 0
-        talk = Talk.objects.first()
-
-        assert talk.tags.count() == 0
-
-        assert Speaker.objects.get()
-
-        # second proposal
-
-        talk_proposal = {
-            "type": "t_45",
-            "title": "More about EPCON testing",
-            "abstract_short": "Longer talk about testing",
-            "abstract": "Using django TestCase and pytest",
-            "level": "advanced",
-            "tags": "django, testing, slides",
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        assert ConferenceTag.objects.count() == 0
-        assert Tag.objects.count() == 0
-
-    def test_ignores_new_tags_keeping_predefined_ones(self):
-        ConferenceTag.objects.create(name='django')
-        ConferenceTag.objects.create(name='love')
-
-        assert Talk.objects.all().count() == 0
-        assert ConferenceTag.objects.count() == 2
-        assert Tag.objects.count() == 0
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        VALIDATION_SUCCESSFUL_302 = 302
-
-        talk_proposal = {
-            "type": "t_30",
-            'first_name': 'Joe',
-            'last_name': 'Doe',
-            "birthday": "2018-02-26",
-            'bio': "Python developer",
-            "title": "Testing EPCON CFP",
-            "abstract_short": "Short talk about testing CFP",
-            "abstract": "Using django TestCase and pytest",
-            "level": "advanced",
-            "phone": "41331237",
-            "tags": "django, testing, slides",
-            "personal_agreement": True,
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        assert ConferenceTag.objects.count() == 2
-        talk = Talk.objects.last()
-
-        assert talk.tags.count() == 1
-
-        assert 'django' in talk.tags.all().values_list('name', flat=True)
-
-        # second proposal
-
-        talk_proposal = {
-            "type": "t_45",
-            "title": "More about EPCON testing",
-            "abstract_short": "Longer talk about testing",
-            "abstract": "Using django TestCase and pytest",
-            "level": "advanced",
-            "tags": "love, testing, slides",
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        assert ConferenceTag.objects.count() == 2
-
-        talk = Talk.objects.first()
-
-        assert talk.tags.count() == 1
-
-        assert talk.title == 'More about EPCON testing'
-        assert 'love' in talk.tags.all().values_list('name', flat=True)
-
-    def test_allows_very_long_description(self):
-        assert Tag.objects.count() == 0
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        VALIDATION_SUCCESSFUL_302 = 302
-
-        abstract = 'a' * 5000
-
-        talk_proposal = {
-            "type": "t_30",
-            'first_name': 'Joe',
-            'last_name': 'Doe',
-            "birthday": "2018-02-26",
-            'bio': "Python developer",
-            "title": "Testing EPCON CFP",
-            "abstract_short": "Short talk about testing CFP",
-            "abstract": abstract,
-            "level": "advanced",
-            "phone": "41331237",
-            "tags": "django, testing, slides",
-            "personal_agreement": True,
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        talk = Talk.objects.first()
-
-        assert talk.abstracts.all()[0].body == abstract
-
-        # second proposal
-
-        talk_proposal = {
-            "type": "t_45",
-            "title": "More about EPCON testing",
-            "abstract_short": "Longer talk about testing",
-            "abstract": abstract,
-            "level": "advanced",
-            "tags": "love, testing, slides",
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        talk = Talk.objects.first()
-
-        assert talk.abstracts.all()[0].body == abstract
-
-    def test_allows_talk_domain_and_domain_level(self):
-        assert Tag.objects.count() == 0
-        self.client.login(email='joedoe@example.com', password='password123')
-
-        VALIDATION_SUCCESSFUL_302 = 302
-        abstract = 'aaaaaaaaaaaa'
-
-        talk_proposal = {
-            "type": "t_30",
-            'first_name': 'Joe',
-            'last_name': 'Doe',
-            "birthday": "2018-02-26",
-            'bio': "Python developer",
-            "title": "Testing EPCON CFP",
-            "abstract_short": "Short talk about testing CFP",
-            "abstract": abstract,
-            "level": TALK_LEVEL.advanced,
-            "phone": "41331237",
-            "domain": "django",
-            "domain_level": TALK_LEVEL.intermediate,
-            "tags": "django, testing, slides",
-            "personal_agreement": True,
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        talk = Talk.objects.first()
-        # Checking issue #654 – if talk is accessible after submission
-        response = self.client.get(talk.get_absolute_url())
-        self.assertTemplateUsed(response, "conference/talk.html")
-
-        assert talk.domain == 'django'
-        assert talk.domain_level == TALK_LEVEL.intermediate
-        assert talk.level == TALK_LEVEL.advanced
-
-        # second proposal
-
-        talk_proposal = {
-            "type": "t_45",
-            "title": "More about EPCON testing",
-            "abstract_short": "Longer talk about testing",
-            "abstract": abstract,
-            "level": TALK_LEVEL.intermediate,
-            "tags": "love, testing, slides",
-            "domain": "devops",
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() + timedelta(days=1),
+    )
+    step1_url = reverse("cfp:step1_submit_proposal")
+    response = user_client.post(step1_url, {})
+    # POST rerenders the same template
+    assert template_used(response, "ep19/bs/cfp/step1_talk_info.html")
+
+
+def test_if_user_can_submit_talk_details_and_is_redirect_to_step2(user_client):
+    STEP1_CORRECT_REDIRECT_302 = 302
+
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() + timedelta(days=1),
+    )
+    step1_url = reverse("cfp:step1_submit_proposal")
+
+    response = user_client.post(
+        step1_url,
+        {
+            "type": TALK_TYPE_CHOICES.t_30,
+            "abstract": "Abstract goes here",
+            "title": "A title",
+            "sub_title": "A sub title",
+            "abstract_short": "Short abstract",
+            "abstract_extra": "Abstract _extra",
+            "tags": "abc, defg",
+            "level": TALK_LEVEL.beginner,
             "domain_level": TALK_LEVEL.advanced,
-            "slides_agreement": True,
-            "video_agreement": True,
+        },
+    )
+
+    assert response.status_code == STEP1_CORRECT_REDIRECT_302
+
+    talk = Talk.objects.get()
+    talk_dict = dump_relevant_talk_information_to_dict(talk)
+    assert talk_dict["type"] == TALK_TYPE_CHOICES.t_30
+    assert talk_dict["type_display"] == "Talk (30 mins)"
+    assert talk_dict["subtitle"] == "A sub title"
+    assert talk_dict["abstract"] == "Abstract goes here"
+    assert talk_dict["abstract_short"] == "Short abstract"
+    assert talk_dict["abstract_extra"] == "Abstract _extra"
+    assert talk_dict["python_level"] == "Beginner"
+    assert talk_dict["domain_level"] == "Advanced"
+    assert talk_dict["speakers"] == []
+
+    assert redirects_to(
+        response, reverse("cfp:step2_add_speakers", args=[talk.uuid])
+    )
+
+
+def test_validation_errors_are_handled_on_step2(user_client):
+    """
+    NOTE(artcz)
+    This test is basically a placholder to get a proper branch coverage.
+    We should also (separately?) test if the validation itslef is correct.
+    """
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() + timedelta(days=1),
+    )
+    talk = TalkFactory()
+    step2_url = reverse("cfp:step2_add_speakers", args=[talk.uuid])
+    response = user_client.post(step2_url, {})
+    # POST rerenders the same template
+    assert template_used(response, "ep19/bs/cfp/step2_add_speaker.html")
+
+
+def test_if_user_can_add_a_speaker_to_a_proposal(user_client):
+    create_conference_with_open_cfp()
+    STEP2_CORRECT_REDIRECT_302 = 302
+
+    talk = TalkFactory()
+    talk.setAbstract(
+        "Setting abstract just because we'll need it to dump later"
+    )
+    step2_url = reverse("cfp:step2_add_speakers", args=[talk.uuid])
+
+    response = user_client.get(step2_url)
+    assert template_used(response, "ep19/bs/cfp/step2_add_speaker.html")
+
+    response = user_client.post(
+        step2_url,
+        {
+            "users_given_name": "Joe Doe",
+            "phone": "+48523456789",
+            "gender": "M",
+            "bio": "ASdf bio",
+        },
+    )
+    assert response.status_code == STEP2_CORRECT_REDIRECT_302
+
+    talk.refresh_from_db()
+    talk_dict = dump_relevant_talk_information_to_dict(talk)
+    speaker = talk_dict["speakers"][0]
+    assert speaker["name"] == "Joe Doe"
+    assert speaker["company"] == ""
+    assert speaker["company_homepage"] == ""
+    assert speaker["bio"] == "ASdf bio"
+    assert speaker["phone"] == "+48523456789"
+
+    assert redirects_to(
+        response, reverse("cfp:step3_thanks", args=[talk.uuid])
+    )
+
+
+def test_if_correct_thanks_page_is_rendered(user_client):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    step3_url = reverse("cfp:step3_thanks", args=[talk.uuid])
+
+    response = user_client.get(step3_url)
+    assert template_used(response, "ep19/bs/cfp/step3_thanks.html")
+
+
+def test_thanks_page_contains_link_to_preview_proposal(user_client):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    step3_url = reverse("cfp:step3_thanks", args=[talk.uuid])
+
+    preview_link = reverse("cfp:preview", args=[talk.slug])
+    response = user_client.get(step3_url)
+
+    assert preview_link in response.content.decode()
+
+
+def test_preview_page_renders_correct_content(user_client):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("Some abstract")
+    preview_url = reverse("cfp:preview", args=[talk.slug])
+
+    response = user_client.get(preview_url)
+    assert talk.title in response.content.decode()
+    assert talk.sub_title in response.content.decode()
+    assert talk.getAbstract().body in response.content.decode()
+    assert talk.abstract_short in response.content.decode()
+    assert talk.abstract_extra in response.content.decode()
+
+
+def test_preview_page_contains_edit_links_if_cfp_is_open_and_user_is_author(
+    user_client
+):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    talk.created_by = user_client.user
+    talk.save()
+    preview_url = reverse("cfp:preview", args=[talk.slug])
+    edit_link = reverse("cfp:update", args=[talk.uuid])
+    speaker_edit_link = reverse("cfp:update_speakers", args=[talk.uuid])
+
+    response = user_client.get(preview_url)
+
+    assert edit_link in response.content.decode()
+    assert speaker_edit_link in response.content.decode()
+
+
+def test_preview_page_doesnt_contain_edit_link_if_cfp_is_open_but_user_is_not_author(  # NOQA
+    user_client
+):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    preview_url = reverse("cfp:preview", args=[talk.slug])
+    edit_link = reverse("cfp:update", args=[talk.uuid])
+    speaker_edit_link = reverse("cfp:update_speakers", args=[talk.uuid])
+
+    response = user_client.get(preview_url)
+
+    assert edit_link not in response.content.decode()
+    assert speaker_edit_link not in response.content.decode()
+
+
+def test_preview_page_doesnt_contain_edit_link_if_cfp_is_closed_and_user_is_author(  # NOQA
+    user_client
+):
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() - timedelta(days=1),
+    )
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    preview_url = reverse("cfp:preview", args=[talk.slug])
+    edit_link = reverse("cfp:update", args=[talk.uuid])
+    speaker_edit_link = reverse("cfp:update_speakers", args=[talk.uuid])
+
+    response = user_client.get(preview_url)
+
+    assert edit_link not in response.content.decode()
+    assert speaker_edit_link not in response.content.decode()
+
+
+def test_regular_user_cant_access_program_wg_download(user_client):
+    LOGIN_REQUIRED_302 = 302
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+    )
+    url = reverse("cfp:program_wg_download_all_talks")
+    response = user_client.get(url)
+    assert response.status_code == LOGIN_REQUIRED_302
+    assert redirects_to(response, "/admin/login/")
+
+
+def test_admin_user_can_access_program_wg_download(admin_client):
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+    )
+    url = reverse("cfp:program_wg_download_all_talks")
+    response = admin_client.get(url)
+    assert response.status_code == 200
+    assert response.json() == {"talks": []}
+
+
+@mark.parametrize(
+    "page_name, template",
+    [
+        ("cfp:update", "ep19/bs/cfp/update_proposal.html"),
+        ("cfp:update_speakers", "ep19/bs/cfp/update_speakers.html"),
+    ],
+)
+def test_update_pages_work_if_cfp_is_open_and_user_is_author(
+    user_client,
+    page_name,
+    template
+):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    talk.created_by = user_client.user
+    talk.save()
+    edit_url = reverse(page_name, args=[talk.uuid])
+
+    response = user_client.get(edit_url)
+
+    assert response.status_code == 200
+    assert template_used(response, template)
+
+
+@mark.parametrize(
+    "page_name",
+    [
+        "cfp:update",
+        "cfp:update_speakers",
+    ],
+)
+def test_update_pages_dont_work_if_cfp_is_open_but_user_is_not_author(
+    user_client, page_name,
+):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    edit_url = reverse(page_name, args=[talk.uuid])
+
+    response = user_client.get(edit_url)
+    assert response.status_code == 403
+
+
+def test_preview_page_doesnt_work_if_cfp_is_closed_and_user_is_author(
+    user_client
+):
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() - timedelta(days=1),
+    )
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    talk.created_by = user_client.user
+    talk.save()
+    edit_url = reverse("cfp:update", args=[talk.uuid])
+
+    response = user_client.get(edit_url)
+    assert response.status_code == 403
+
+
+def test_validation_errors_are_handled_on_update_proposal(user_client):
+    """
+    NOTE(artcz)
+    This test is basically a placholder to get a proper branch coverage.
+    We should also (separately?) test if the validation itslef is correct.
+    """
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() + timedelta(days=1),
+    )
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    talk.created_by = user_client.user
+    talk.save()
+    update_url = reverse("cfp:update", args=[talk.uuid])
+    response = user_client.post(update_url, {})
+
+    # POST rerenders the same template
+    assert template_used(response, "ep19/bs/cfp/update_proposal.html")
+
+
+def test_update_proposal_updates_proposal(user_client):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    talk.created_by = user_client.user
+    talk.save()
+
+    edit_url = reverse("cfp:update", args=[talk.uuid])
+
+    response = user_client.post(
+        edit_url,
+        {
+            "type": TALK_TYPE_CHOICES.t_45,
+            "abstract": "New abstract",
+            "abstract_short": "New short abstract",
+            "abstract_extra": "New extra abstract",
+            "level": TALK_LEVEL.intermediate,
+            "domain_level": TALK_LEVEL.advanced,
+            "title": "New title",
+            "sub_title": "New sub title",
+            "tags": "Some, tags",
+        },
+    )
+
+    assert response.status_code == 302
+    talk.refresh_from_db()
+
+    assert redirects_to(response, reverse("cfp:preview", args=[talk.slug]))
+
+    talk_dict = dump_relevant_talk_information_to_dict(talk)
+    assert talk_dict["type"] == TALK_TYPE_CHOICES.t_45
+    assert talk_dict["type_display"] == "Talk (45 mins)"
+    assert talk_dict["subtitle"] == "New sub title"
+    assert talk_dict["abstract"] == "New abstract"
+    assert talk_dict["abstract_short"] == "New short abstract"
+    assert talk_dict["abstract_extra"] == "New extra abstract"
+    assert talk_dict["python_level"] == "Intermediate"
+    assert talk_dict["domain_level"] == "Advanced"
+    assert talk_dict["speakers"] == []
+
+
+def test_update_speaker_updated_speaker(user_client):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    talk.created_by = user_client.user
+    talk.save()
+
+    edit_url = reverse("cfp:update_speakers", args=[talk.uuid])
+    speaker_data = dict(
+        users_given_name="new name",
+        is_minor=True,
+        job_title="goat",
+        phone="+48123456789",
+        gender="yes",
+        company="widgets inc",
+        company_homepage="www.widgets.inc",
+        bio="this is my bio",
+    )
+
+    response = user_client.post(edit_url, speaker_data)
+
+    assert response.status_code == 302
+    assert redirects_to(response, reverse("cfp:preview", args=[talk.slug]))
+
+    user = user_client.user
+    user.refresh_from_db()
+    attendee_profile = user.attendeeprofile
+    attendee_profile.refresh_from_db()
+    assert user.assopy_user.name() == speaker_data["users_given_name"]
+    assert attendee_profile.phone == speaker_data["phone"]
+    assert attendee_profile.is_minor == speaker_data["is_minor"]
+    assert attendee_profile.job_title == speaker_data["job_title"]
+    assert attendee_profile.gender == speaker_data["gender"]
+    assert attendee_profile.company == speaker_data["company"]
+    assert speaker_data["company_homepage"] in attendee_profile.company_homepage
+    assert attendee_profile.getBio().body == speaker_data["bio"]
+
+
+def test_update_speaker_updated_speaker_name(user_client):
+    create_conference_with_open_cfp()
+    talk = TalkFactory()
+    talk.setAbstract("some abstract")
+    talk.created_by = user_client.user
+    talk.save()
+
+    user = user_client.user
+    user.first_name = "John"
+    user.last_name = "Doe"
+    user.save()
+
+    edit_url = reverse("cfp:update_speakers", args=[talk.uuid])
+    speaker_data = dict(
+        users_given_name="new name",
+        phone="+48123456789",
+        bio="this is my bio",
+    )
+
+    response = user_client.post(edit_url, speaker_data)
+
+    assert response.status_code == 302
+    assert redirects_to(response, reverse("cfp:preview", args=[talk.slug]))
+
+    user.refresh_from_db()
+    attendee_profile = user.attendeeprofile
+    attendee_profile.refresh_from_db()
+    assert user.assopy_user.name() == speaker_data["users_given_name"]
+
+
+# Mark with django db only because AddSpeakerToTalkForm is a ModelForm
+@mark.parametrize(
+    "valid_phone", ["+48123456789", "+44 7 123 456 789", "+1 858 712 8966"]
+)
+def test_speaker_form_accepts_valid_international_mobile_numbers(valid_phone):
+    form = AddSpeakerToTalkForm(
+        {
+            "users_given_name": "Joe Doe",
+            "phone": valid_phone,
+            "gender": "",
+            "bio": "ASdf bio",
         }
+    )
+    assert form.is_valid()
 
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
 
-        talk = Talk.objects.first()
-        # Checking issue #654 – if talk is accessible after submission
-        response = self.client.get(talk.get_absolute_url())
-        self.assertTemplateUsed(response, "conference/talk.html")
-
-        assert talk.abstracts.all()[0].body == abstract
-        assert talk.domain == 'devops'
-        assert talk.domain_level == TALK_LEVEL.advanced
-        assert talk.level == TALK_LEVEL.intermediate
-
-    def test_662_editing_cfp_proposal(self):
-        """
-        https://github.com/EuroPython/epcon/issues/662
-        """
-
-        self.client.login(email=self.user.email, password='password123')
-        VALIDATION_SUCCESSFUL_302 = 302
-        EDIT_SUCCESSFUL_303 = 303
-        abstract = 'aaaaaaaaaaaa'
-
-        talk_proposal = {
-            "type": "t_30",
-            'first_name': 'Joe',
-            'last_name': 'Doe',
-            "birthday": "2018-02-26",
-            'bio': "Python developer",
-            "title": "Testing EPCON CFP",
-            "abstract_short": "Short talk about testing CFP",
-            "abstract": abstract,
-            "level": TALK_LEVEL.advanced,
-            "phone": "41331237",
-            "domain": "django",
-            "domain_level": TALK_LEVEL.intermediate,
-            "tags": "django, testing, slides",
-            "personal_agreement": True,
-            "slides_agreement": True,
-            "video_agreement": True,
+# Mark with django db only because AddSpeakerToTalkForm is a ModelForm
+@mark.parametrize(
+    "invalid_phone",
+    ["SOME RANODM TEXT", "+4471343956789", "+4412445367"],
+)
+def test_speaker_form_doesnt_accept_invalid_international_mobile_numbers(
+    invalid_phone
+):
+    form = AddSpeakerToTalkForm(
+        {
+            "users_given_name": "Joe Doe",
+            "phone": invalid_phone,
+            "gender": "",
+            "bio": "ASdf bio",
         }
+    )
+    assert not form.is_valid()
+    assert (
+        form.errors["phone"]
+        == ["Enter a valid phone number (e.g. +12125552368)."]
+    )
 
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
 
-        talk = Talk.objects.first()
-        assert talk.abstract_short == "Short talk about testing CFP"
-        self.assertTemplateUsed(
-            self.client.get(talk.get_absolute_url()),
-            "conference/talk.html"
-        )
-
-        talk_proposal["abstract_short"] = "First edit"
-        response = self.client.post(talk.get_absolute_url(), talk_proposal)
-        assert response.status_code == EDIT_SUCCESSFUL_303
-
-        talk = Talk.objects.first()
-        assert talk.abstract_short == "First edit"
-
-        talk_proposal["abstract_short"] = "Second edit"
-        response = self.client.post(talk.get_absolute_url(), talk_proposal)
-        assert response.status_code == EDIT_SUCCESSFUL_303
-
-        talk = Talk.objects.first()
-        assert talk.abstract_short == "Second edit"
-
-    def test_665_edit_with_other_in_django_admin(self):
-        """
-        https://github.com/EuroPython/epcon/issues/665
-        """
-        self.client.login(email=self.user.email, password='password123')
-        VALIDATION_SUCCESSFUL_302 = 302
-        EDIT_SUCCESSFUL_302       = 302
-        abstract = 'aaaaaaaaaaaa'
-
-        talk_proposal = {
-            "type": "t_30",
-            'first_name': 'Joe',
-            'last_name': 'Doe',
-            "birthday": "2018-02-26",
-            'bio': "Python developer",
-            "title": "Testing EPCON CFP",
-            "abstract_short": "Short talk about testing CFP",
-            "abstract": abstract,
-            "level": TALK_LEVEL.advanced,
-            "phone": "41331237",
-            "domain": settings.CONFERENCE_TALK_DOMAIN.django,
-            "domain_level": TALK_LEVEL.intermediate,
-            "tags": "django, testing, slides",
-            "personal_agreement": True,
-            "slides_agreement": True,
-            "video_agreement": True,
-        }
-
-        response = self.client.post(self.form_url, talk_proposal)
-        assert response.status_code == VALIDATION_SUCCESSFUL_302
-
-        talk = Talk.objects.first()
-
-        # relogin as staff user
-        self.client.login(email='admin@example.com', password='admin')
-        response = self.client.get(talk.get_admin_url())
-        self.assertTemplateUsed(response, "admin/change_form.html")
-
-        # TODO: this should be probably dynamically generated from the GET
-        # response above, but not sure how to correctly dump formset to dict,
-        # so instead I just served it's response and did parse_qsl on the
-        # whole POST.
-        admin_talk_edit = {
-            '_save': 'Save',
-            'abstract_short': 'Short talk about testing CFP',
-            'abstracts_en': 'aaaaaaaaaaaa',
-            'conference': 'ep2018',
-            'domain': settings.CONFERENCE_TALK_DOMAIN.other,
-            'domain_level': 'intermediate',
-            'duration': '30',
-            'language': 'en',
-            'level': 'advanced',
-            'slug': 'testing-epcon-cfp',
-            'status': 'proposed',
-            'tags': "django, testing, slides",
-            'talkspeaker_set-0-id': '1',
-            'talkspeaker_set-0-speaker': '2',
-            'talkspeaker_set-0-talk': '1',
-            'talkspeaker_set-1-talk': '1',
-            'talkspeaker_set-INITIAL_FORMS': '1',
-            'talkspeaker_set-MAX_NUM_FORMS': '1000',
-            'talkspeaker_set-MIN_NUM_FORMS': '0',
-            'talkspeaker_set-TOTAL_FORMS': '2',
-            'talkspeaker_set-__prefix__-talk': '1',
-            'title': 'Testing EPCON CFP',
-            'type': 't_30'
-        }
-
-        assert settings.CONFERENCE_TALK_DOMAIN.other == ''
-        response = self.client.post(talk.get_admin_url(), admin_talk_edit)
-        assert response.status_code == EDIT_SUCCESSFUL_302
-
-        talk = Talk.objects.first()
-        assert talk.domain != settings.CONFERENCE_TALK_DOMAIN.django
-        assert talk.domain == settings.CONFERENCE_TALK_DOMAIN.other
-        assert talk.domain == ''
+def create_conference_with_open_cfp():
+    Conference.objects.create(
+        code=settings.CONFERENCE_CONFERENCE,
+        name=settings.CONFERENCE_CONFERENCE,
+        cfp_start=timezone.now().date() - timedelta(days=2),
+        cfp_end=timezone.now().date() + timedelta(days=1),
+    )
